+++ xml/src/main/scala/org/podval/xml/XmlAttribute.scala
@@ -5,6 +5,7 @@ open class XmlAttribute(val name: String)
 object XmlAttribute:
   object Id extends XmlAttribute("id")
   object XmlId extends XmlAttribute("xml:id")
+  object XmlBase extends XmlAttribute("xml:base")
 
   object Xmlns extends XmlAttribute("xmlns"):
     def apply(ns: String): XmlAttribute = XmlAttribute(s"$name:$ns")